docs: record where Sonic Frontiers' mip levels are, and two apparatus traps
- SONIC_FRONTIERS_STATUS.md: correct "All three now block on exactly one
  instruction" (they did not), record the two-blocker A/B and its method, name
  the declining resource (a 2048x2048 RG32F 12-level pyramid the guest builds
  itself, thirteen T#s over one allocation), and point whoever takes #2818 at
  the fact that the levels are ordinary guest memory whose offsets prosper
  already computes.
- Three new `## Ruled out` rows: the "single remaining blocker" claim,
  `--warmup-seconds` desynchronising the flip-anchored live route, and the
  routed census desynchronising under host load. The last one is why the
  reproduction section now says to read the program count on the census line
  before believing anything under it.
- RECOMPILER_REMAINING.md: a FIXED section for the typed embedded-table fold,
  because the gap is cross-title and the reject it produces names the getpc
  rather than the load that carries the information.
- BLOG.md: an entry for the finding.
- Print the two `bool` MIMG modifier fields through explicit `unsigned` casts.

Refs #2859, #2818, #2790, #1891.

+++ prosper/src/gpu/recompiler/rdna2_emit_alu.cpp
@@ -7189,7 +7189,8 @@ bool emit_alu(SpirvCompute& b, RegState& rs, const Rdna2Inst& in, bool& ok, bool
                                  (int)(in.mimg_dim == 5u && !b.is_compute),
                                  (unsigned long long)res->gpu_addr, res->width, res->height,
                                  res->depth, (int)res->format, res->tile_mode, in.mimg_dmask,
-                                 in.mimg_unorm, in.mimg_glc, res->layer_stride_bytes);
+                                 (unsigned)in.mimg_unorm, (unsigned)in.mimg_glc,
+                                 res->layer_stride_bytes);
                 ok = false;
                 return true;
             }
